% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_leaves_as_df}
\alias{get_leaves_as_df}
\title{Given a DAG return a dataframe with the leaves and parents
of a node vector}
\usage{
get_leaves_as_df(dag, nodes)
}
\description{
Given a DAG return a dataframe with the leaves and parents
of a node vector
}
